# openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "RealAdvisor Inbox API",
		"version": "1.0.0",
...
	},
	"servers": [
		{
			"url": "/",
			"description": "Current inbox instance"
		}
	],
	"security": [
		{
			"AccessSession": []
		}
	],
...
	"paths": {
...
		"/api/v1/mailboxes/{mailboxId}/agent": {
			"get": {
				"operationId": "getAgentState",
				"tags": ["Agent"],
				"summary": "Get agent conversation",
				"parameters": [
					{
						"name": "mailboxId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "conversationId",
						"in": "query",
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "before",
						"in": "query",
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Availability, settings, and 30 most recent turns with model, status and saved tool actions",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"available": {
											"type": "boolean"
										},
										"autoDraftAvailable": {
											"type": "boolean"
										},
										"settings": {
											"type": "object",
											"required": ["model", "system_prompt", "auto_draft"],
											"additionalProperties": false,
											"properties": {
												"model": {
													"type": "string",
													"maxLength": 200,
													"description": "Available model ID from the model catalog. Workers AI IDs start with @cf/; gateway IDs use provider/model."
												},
												"system_prompt": {
													"type": "string",
													"maxLength": 8000
												},
												"auto_draft": {
													"type": "boolean"
												}
											}
										},
										"turns": {
											"type": "array",
											"items": {
												"type": "object"
											}
										},
										"catalog": {
											"$ref": "#/components/schemas/AgentCatalog"
										},
										"conversation": {
											"type": "object"
										},
										"hasMore": {
											"type": "boolean"
										}
									}
								}
							}
						}
					}
				},
				"security": [
					{
						"AccessSession": []
					},
					{
						"InboxApiKey": []
					}
				],
				"description": "Read agent settings and a page of one conversation.\n\nAPI keys require agent:use and mail:read and access to this mailbox."
			}
		},
		"/api/v1/mailboxes/{mailboxId}/agent/settings": {
			"put": {
				"operationId": "putAgentSettings",
				"tags": ["Agent"],
				"summary": "Update agent settings",
				"parameters": [
					{
						"name": "mailboxId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": ["model", "system_prompt", "auto_draft"],
								"additionalProperties": false,
								"properties": {
									"model": {
										"type": "string",
										"maxLength": 200,
										"description": "Available model ID from the model catalog. Workers AI IDs start with @cf/; gateway IDs use provider/model."
									},
									"system_prompt": {
										"type": "string",
										"maxLength": 8000
									},
									"auto_draft": {
										"type": "boolean"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Saved settings",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": ["model", "system_prompt", "auto_draft"],
									"additionalProperties": false,
									"properties": {
										"model": {
											"type": "string",
											"maxLength": 200,
											"description": "Available model ID from the model catalog. Workers AI IDs start with @cf/; gateway IDs use provider/model."
										},
										"system_prompt": {
											"type": "string",
											"maxLength": 8000
										},
										"auto_draft": {
											"type": "boolean"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Invalid model or settings"
					},
					"503": {
						"description": "Selected provider is not configured, or automatic drafting queue is unavailable"
					}
				},
				"description": "Choose the default model, writing instructions, and automatic drafting for this mailbox."
			}
		},
		"/api/v1/mailboxes/{mailboxId}/agent/chat": {
			"post": {
				"operationId": "postAgentChat",
				"tags": ["Agent"],
				"summary": "Chat with agent",
				"description": "Stream an email agent turn; draft and organization tools can modify this mailbox.\n\nNever sends email. Streams the native Vercel AI SDK UI message protocol (SSE), including text and structured tool inputs/results. Use DefaultChatTransport with id and prompt from the latest user message. The server loads trusted history from Postgres. Duplicate request IDs return 409 and are not replayed. Disconnected runs abort and persist partial results; read history before retrying. API keys require agent:use and mail:read and access to this mailbox.",
				"parameters": [
					{
						"name": "mailboxId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"additionalProperties": false,
								"required": ["id", "prompt"],
								"properties": {
									"id": {
										"type": "string",
										"format": "uuid"
									},
									"prompt": {
										"type": "string",
										"minLength": 1,
										"maxLength": 8000
									},
									"emailId": {
										"type": "string",
										"format": "uuid"
									},
									"model": {
										"type": "string",
										"maxLength": 200,
										"description": "Optional model override for this turn. Does not change the mailbox default."
									},
									"conversationId": {
										"type": "string",
										"format": "uuid",
										"description": "Conversation in the current mailbox; omitted by legacy clients."
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Native Vercel AI SDK UI message stream",
						"headers": {
							"x-vercel-ai-ui-message-stream": {
								"schema": {
									"type": "string",
									"enum": ["v1"]
								}
							},
							"x-chat-model": {
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"text/event-stream": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"400": {
						"description": "Unknown, retired or invalid model, or invalid request"
					},
					"409": {
						"description": "Mailbox is busy or request ID was already used"
					},
					"503": {
						"description": "Selected AI provider is not configured"
					}
				},
				"security": [
					{
						"AccessSession": []
					},
					{
						"InboxApiKey": []
					}
				]
			}
		},
...
		"/api/v1/agent/models/refresh": {
			"post": {
				"operationId": "refreshAgentModels",
				"tags": ["Agent"],
				"summary": "Refresh agent models",
				"responses": {
					"200": {
						"description": "Model catalog and provider availability",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AgentCatalog"
								}
							}
						}
					},
					"409": {
						"description": "Another refresh is in progress"
					},
					"502": {
						"description": "Upstream catalog unavailable or invalid; existing snapshot is unchanged"
					}
				},
				"description": "Refresh gateway models without changing mailbox defaults."
			}
		},
		"/api/v1/mailboxes/{mailboxId}/agent/stop": {
			"post": {
				"operationId": "stopAgentRun",
				"tags": ["Agent"],
				"summary": "Stop agent run",
				"description": "Stop a specific mailbox agent run.\n\nRevokes the run mutation lease before returning. Generation aborts when the worker observes the stop. Existing saved changes are retained. Repeated stops cannot cancel a newer run. API keys require agent:use and mail:read and access to this mailbox.",
				"parameters": [
					{
						"name": "mailboxId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": ["id"],
								"additionalProperties": false,
								"properties": {
									"id": {
										"type": "string",
										"format": "uuid"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Stop acknowledged; stopped is false if the run was already terminal",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": ["stopped"],
									"properties": {
										"stopped": {
											"type": "boolean"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Invalid run ID"
					},
					"404": {
						"description": "Run not found in this mailbox"
					}
				},
				"security": [
					{
						"AccessSession": []
					},
					{
						"InboxApiKey": []
					}
				]
			}
		},
...
		"/api/v1/mailboxes/{mailboxId}/agent/conversations": {
			"get": {
				"operationId": "listAgentConversations",
				"tags": ["Agent"],
				"summary": "List agent conversations",
				"parameters": [
					{
						"name": "mailboxId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "offset",
						"in": "query",
						"schema": {
							"type": "integer",
							"minimum": 0,
							"maximum": 100000
						}
					},
					{
						"name": "search",
						"in": "query",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Up to 50 conversations and hasMore"
					}
				},
				"security": [
					{
						"AccessSession": []
					},
					{
						"InboxApiKey": []
					}
				],
				"description": "Browse saved assistant conversations, newest first.\n\nAPI keys require agent:use and mail:read and access to this mailbox."
			},
			"post": {
				"operationId": "createAgentConversation",
				"tags": ["Agent"],
				"summary": "Create agent conversation",
				"parameters": [
					{
						"name": "mailboxId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"201": {
						"description": "New conversation with id, title, and timestamps"
					}
				},
				"security": [
					{
						"AccessSession": []
					},
					{
						"InboxApiKey": []
					}
				],
				"description": "Start a separate assistant conversation.\n\nAPI keys require agent:use and mail:read and access to this mailbox."
			}
		},
...
		"/api/v1/mailboxes/{mailboxId}/agent/compose": {
			"post": {
				"tags": ["Agent"],
				"summary": "Generate email suggestion",
				"parameters": [
					{
						"name": "mailboxId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"additionalProperties": false,
								"properties": {
									"emailId": {
										"type": "string",
										"format": "uuid"
									},
									"action": {
										"type": "string",
										"enum": ["draft", "improve", "shorten", "formalize"],
										"default": "draft"
									},
									"instructions": {
										"type": "string",
										"maxLength": 2000
									},
									"body": {
										"type": "string",
										"maxLength": 16000
									},
									"subject": {
										"type": "string",
										"maxLength": 1000
									},
									"recipients": {
										"type": "string",
										"maxLength": 2000
									},
									"model": {
										"type": "string",
										"maxLength": 200,
										"description": "Optional model override from the existing agent model catalog; defaults to the mailbox model"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Plain-text email body and model used",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": ["text", "model"],
									"properties": {
										"text": {
											"type": "string"
										},
										"model": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Invalid input or unavailable model"
					},
					"404": {
						"description": "Mailbox or source email not found"
					},
					"502": {
						"description": "Model generation failed"
					},
					"503": {
						"description": "Provider not configured"
					},
					"504": {
						"description": "Generation cancelled or timed out"
					}
				},
				"security": [
					{
						"AccessSession": []
					},
					{
						"InboxApiKey": []
					}
				],
				"description": "Generate an email body suggestion without saving or sending.\n\nAPI keys require agent:use and mail:read and access to this mailbox."
			},
			"get": {
				"tags": ["Agent"],
				"summary": "Get writing settings",
				"parameters": [
					{
						"name": "mailboxId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Mailbox settings and provider-aware agent model catalog"
					},
					"404": {
						"description": "Mailbox not found"
					}
				},
				"security": [
					{
						"AccessSession": []
					},
					{
						"InboxApiKey": []
					}
				],
				"description": "Get mailbox writing settings and the shared agent model catalog.\n\nAPI keys require agent:use and mail:read and access to this mailbox."
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"AgentCatalog": {
				"type": "object",
				"required": ["models", "refreshed_at", "gatewayConfigured"],
				"properties": {
					"refreshed_at": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"gatewayConfigured": {
						"type": "boolean"
					},
					"models": {
						"type": "array",
						"items": {
							"type": "object",
							"required": [
								"id",
								"name",
								"provider",
								"source",
								"available",
								"selectable"
							],
							"properties": {
								"id": {
									"type": "string",
									"maxLength": 200,
									"description": "Available model ID from the model catalog. Workers AI IDs start with @cf/; gateway IDs use provider/model."
								},
								"name": {
									"type": "string"
								},
								"provider": {
									"type": "string"
								},
								"source": {
									"type": "string",
									"enum": ["workers", "gateway"]
								},
								"available": {
									"type": "boolean",
									"description": "Present in the latest catalog"
								},
								"selectable": {
									"type": "boolean",
									"description": "Available and its provider is configured"
								},
								"context_window": {
									"type": "integer",
									"nullable": true
								},
								"input_price": {
									"type": "number",
									"nullable": true,
									"description": "USD per million input tokens"
								},
								"output_price": {
									"type": "number",
									"nullable": true,
									"description": "USD per million output tokens"
								}
							}
						}
					}
				}
			},
...
		},
...
		"securitySchemes": {
			"AccessSession": {
				"type": "apiKey",
				"in": "cookie",
				"name": "CF_Authorization",
				"description": "Verified Cloudflare Access browser session; human membership and roles apply. Do not copy session cookies to integrations."
			},
			"InboxApiKey": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "inbox_<64 hex characters>",
				"description": "Created by an administrator in Settings \u2192 API keys. Scope and mailbox restrictions apply; revoked or expired keys return 401."
			}
		}
	}
}